Load wp-load.php dynamically (updated)
I see that you are planning to fix this in an upcoming patch…in case
you are still going to require `wp-load.php` I wanted to suggest the
following fix.

I am only submitting this again because I changed it to use
`$_SERVER['SCRIPT_FILENAME’]` instead of `__FILE__`. The reason for
this is that `__FILE__` automatically resolves symlinks. This is a
problem when you are symlinking a custom wp-content folder which lives
outside of the web root folder (i.e. when using composer to manage WP
updates).

Example folder structure where this is applicable:

├── content
│   ├── mu-plugins
│   ├── plugins
│   ├── themes
│   └── uploads
├── composer.json
├── public
│   ├── @content (symlinks to ../content which is a custom wp-content
folder)
│   ├── wp-admin
│   ├── wp-content
│   ├── wp-includes
│   └── etc.
├── vendor
└── wp-config.php