Update class-fields.php
When adding custom social fields, the documentation shows it's possible to set a "url_target" for the field. For my personal use I wanted to have it set to "_self" but it never was applied.
Today when I was trying to figure out a shortcode to display the social fields anywhere, I found the function that displays the social fields and while reading through it, I noticed that the target attribute of the anchor element wasn't reading the "url_target" item and always set it to "_blank".
I added a ternary operator to either use the set "url_target" or fallback to the default "_blank".